reformatted the code

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,4 +1,5 @@
 #include "search_algos.h"
+
 /**
  * linear_search - Search for a value in an array of integers
  * @array: Element of the array to search
@@ -8,13 +9,13 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 
 	if (array == NULL)
 		return (-1);
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int)size; i++)
 	{
-		printf("Value checked array[%lu] = [%i]\n", i, array[i]);
+		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
 			return (i);
